ML model/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
pyyaml>=6.0.0
